Add hovering effect to the cells
Create mousedown and mousemove event listeners and attach an
addHovering handler to them to create a hover effect.
Add a mousedown event listener to the window to remove the
addHovering handler. This prevents the container from being hovered
when the mouse pointer moves out if it.
Implement a createCell function to improve code readability.
In the style.css, add .hover class and include user-select: none to the
.cell class. This prevents bugs with event listeners.