Implement ability to choose cells count
In the index.html, add a button with the id "set-cells-amount".

In the app.js, move the code for creating the game field into a
function called createGameField to reuse this code multiple times.

Assign the button mentioned above to a setCellsAmount variable and
attach a click event listener to it. This will allow capturing user
input and creating a new game field with a specified number of cells.

Add a resize event listener to the window to recalculate the cells
width when the window size changes.

Create a function called removeCells and add it to the createGameField
function to clear the field before appending new cells.

Add an icon to the document.